requirements.txt
flask
flask_wtf
pandas
requests
mysql-connector-python
flask_mail
Werkzeug==2.2.2
itsdangerous
secure-smtplib
bcrypt
pandas
stripe
PrettyTable
datetime
xmltodict
python_random_strings
openai==0.28
PyPDF2
docx2txt
flask_cors
PyJWT
flask_httpauth
python-dotenv
